Updated Activity Item Structure
