+++ App.js
@@ -1,20 +1,13 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import { createStore, applyMiddleware } from 'redux';
-import axios from 'axios'
-import axiosMiddleware from 'redux-axios-middleware';
 import thunk from 'redux-thunk';
 
 import rootReducer from './store/reducers/rootReducer';
-import App from './screens/Main';
+import Main from './screens/Main';
 
-const client = axios.create({});
-const store = createStore(rootReducer, applyMiddleware(thunk, axiosMiddleware(client)));
+const store = createStore(rootReducer, applyMiddleware(thunk));
 
-const ReduxApp = () => {
-  return (
-    <Provider store={store} ><App /></Provider>
-  )
-};
+const App = () => <Provider store={store} ><Main /></Provider>
 
-export default ReduxApp;
+export default App;
